arm: avoid cps instructions where ineffective
Because it serializes (regardless of whether it has any effect), the CPS
instruction is expensive.  Where possible, avoid it.  This also means
not restoring FIQ state (or anything else in CPSR), so it's probably not
portable.

Bonus hack: remove asm clobbers on arch_local_irq... functions, since
they shouldn't be needed.